Moved uploadImage endpoint to Image model
